Created game's base
Have most of the assets that will be used.
Has a game board made with columns for cards, spots to stack the cards to win, and a deck.
The game is able to generate all the cards, shuffle them, and deal them out.
The deck can be clicked on to flip over 3 cards at a time and once cycled through, restart from the beginning.