Started making card dragging
Added an icon for when the deck is empty.
Made a dectection for when a card is being clicked and if the click is being held.
Modified the card to have multiple hit box's for different amounts of detection.
Made basic dragging for cards.
Made some mechanics for locking the cards to a column when released.