Fixed dragging bug 2
Fixed a bug where when a card was dragged outside of a hitbox and released, it would crash.
Added function get_top_parent to card_column and card to help with the above bug.
Moved where child_card_count is initialized when a card is made to guarantee it will always have a value and removed checks for null that are no longer needed elsewhere.

Removed unused function add_extra_cards() in card_column.gd and card.gd.
Reorganized add_card() in card_column.gd to read more smoothly.
Edited comments.