Added logic to the foundation slots
Made a add_card_foundation, change_click_detections_foundation, and check_card_validity_foundation in card.gd for the slightly different logic used when moving cards to the foundation.

Made foundation_card_slot.gd that has all the same functions as card_column.gd but changed as needed to fit the foundation.

Removed unnecessary hitbox changes in on_card_release().
Renamed add_card in card.gd to add_card_column so that it's more clear which object it works with.
Renamed change_click_detections in card.gd to change_click_detections_column so that it's more clear which object it works with.
Renamed check_card_validity in card.gd to check_card_validity_column so that it's more clear which object it works with.