Added undo function
Added an undo function on ctrl + z that undoes your last move.

Changed the flip_card_up() and flip_top_card_up() functions to return true if they did flip a card up and false otherwise.
Added a flip_bottom_card_back() function that flips the bottom most card of a card column to its backside.
Added an undo_deck_flip() function to deck that undoes  a flip3().
Added an array and structure to game_board to keep track of moves.
Added an undo() function to game_board() that takes the last move in the move array and undoes it.